+++ src/ResourceManager/EventHub/ChangeLog.md
@@ -18,6 +18,9 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Updated to the latest version of the Azure ClientRuntime
+
+## Version 0.6.2
 * Fix issue with Default Resource Group in CloudShell
 
 ## Version 0.6.1