+++ src/ResourceManager/Network/ChangeLog.md
@@ -19,8 +19,11 @@
 --->
 
 ## Current Release
-* Fix issue with Default Resource Group in CloudShell
 * Fix error message with Network cmdlets
+* Updated to the latest version of the Azure ClientRuntime
+
+## Version 5.4.1
+* Fix issue with Default Resource Group in CloudShell
 
 ## Version 5.4.0
 * Updating model types for compatibility with DNS cmdlets.