+++ src/ResourceManager/OperationalInsights/ChangeLog.md
@@ -18,6 +18,9 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Updated to the latest version of the Azure ClientRuntime
+
+## Version 4.3.1
 * Fix issue with Default Resource Group in CloudShell
 * Fixed issue with cleaning up scripts in build